Move application Database and Context providers into a utility sub folder for the app directory - Update imports across app
